added fix based on the dev comment
