+++ scripts/batchimagesA.py
@@ -10,14 +10,6 @@ from modules.processing import process_images
 from modules.shared import state
 
 
-def load_prompt_file(file):
-    if file is None:
-        lines = []
-    else:
-        lines = [x.strip() for x in file.decode('utf8', errors='ignore').split("\n")]
-
-    return None, "\n".join(lines), gr.update(lines=7)
-
 
 class Script(scripts.Script):
 
@@ -41,24 +33,7 @@ class Script(scripts.Script):
     # The returned values are passed to the run method as parameters.
 
     def ui(self, is_txt2img):
-
-        script_overrides = gr.CheckboxGroup(label="Overrides", choices=sc.possible_overrides, value=sc.default_overrides)
-        with gr.Accordion(label="Prompt overrides", open=False):
-            prepend_prompt_text = gr.Textbox(label="Text to prepend", lines=1,
-                                             elem_id=self.elem_id("prepend_prompt_text"))
-            append_prompt = gr.Checkbox(label="Append text instead", elem_id=self.elem_id("append_prompt"))
-
-        prompt_txt = gr.Textbox(label="List of prompt inputs", lines=1, elem_id=self.elem_id("prompt_txt"))
-        file = gr.File(label="Upload prompt inputs", type='binary', elem_id=self.elem_id("file"))
-
-        file.change(fn=load_prompt_file, inputs=[file], outputs=[file, prompt_txt, prompt_txt])
-
-        # We start at one line. When the text changes, we jump to seven lines, or two lines if no \n.
-        # We don't shrink back to 1, because that causes the control to ignore [enter], and it may
-        # be unclear to the user that shift-enter is needed.
-        prompt_txt.change(lambda tb: gr.update(lines=7) if ("\n" in tb) else gr.update(lines=2), inputs=[prompt_txt],
-                          outputs=[prompt_txt])
-        return [prepend_prompt_text, append_prompt, prompt_txt, script_overrides]
+        sc.ui(self)
 
     # This is where the additional processing is implemented. The parameters include
     # self, the model object "p" (a StableDiffusionProcessing class, see
