+++ scripts/script_common.py
@@ -1,3 +1,6 @@
+import gradio as gr
+import modules.scripts as scripts
+
 arg_mapping = {
     'Prompt': 'prompt',
     'Negative prompt': 'negative_prompt',
@@ -128,3 +131,31 @@ overrides_mapping = {
 }
 possible_overrides = list(overrides_mapping.keys())
 default_overrides = ["Override Model"]
+
+
+def load_prompt_file(file):
+    if file is None:
+        lines = []
+    else:
+        lines = [x.strip() for x in file.decode('utf8', errors='ignore').split("\n")]
+
+    return None, "\n".join(lines), gr.update(lines=7)
+
+def ui(script: scripts.Script):
+
+    script_overrides = gr.CheckboxGroup(label="Overrides", choices=possible_overrides, value=default_overrides)
+    with gr.Accordion(label="Prompt overrides", open=False):
+        prepend_prompt_text = gr.Textbox(label="Text to prepend", lines=1, elem_id=script.elem_id("prepend_prompt_text"))
+        append_prompt = gr.Checkbox(label="Append text instead", elem_id=script.elem_id("append_prompt"))
+
+    prompt_txt = gr.Textbox(label="List of prompt inputs", lines=1, elem_id=script.elem_id("prompt_txt"))
+    file = gr.File(label="Upload prompt inputs", type='binary', elem_id=script.elem_id("file"))
+
+    file.change(fn=load_prompt_file, inputs=[file], outputs=[file, prompt_txt, prompt_txt])
+
+    # We start at one line. When the text changes, we jump to seven lines, or two lines if no \n.
+    # We don't shrink back to 1, because that causes the control to ignore [enter], and it may
+    # be unclear to the user that shift-enter is needed.
+    prompt_txt.change(lambda tb: gr.update(lines=7) if ("\n" in tb) else gr.update(lines=2), inputs=[prompt_txt],
+                      outputs=[prompt_txt])
+    return [prepend_prompt_text, append_prompt, prompt_txt, script_overrides]
